simple paypal doing amount modifier and moving amount modifier module into scripts dir
